refactor: moved `eval_phase_two` tests
