+++ wallet/src/order_book.rs
@@ -698,3 +698,456 @@ impl From<PallasPlutusData> for OrderDatum {
         }
     }
 }
+
+#[test]
+fn test_phase2_order_start() {
+    use core::str::FromStr;
+    use griffin_core::checks_interface::conway_minted_tx_from_cbor;
+    use griffin_core::h224::H224;
+    use griffin_core::pallas_codec::utils::MaybeIndefArray::Def;
+    use griffin_core::pallas_crypto::hash::Hash;
+    use griffin_core::pallas_primitives::conway::{
+        Constr, MintedTx as ConwayMintedTx, PlutusData as PallasPlutusData,
+    };
+    use griffin_core::types::{
+        compute_plutus_v2_script_hash, Address, AssetClass, AssetName, Datum, ExUnits, Multiasset,
+        Output, PlutusData, PlutusScript, Redeemer, RedeemerTag, VKeyWitness, Value,
+    };
+    use griffin_core::uplc::tx::{eval_phase_two, SlotConfig};
+
+    let script = PlutusScript(hex::decode("59080c010000323232323232322253232323330063001300737540082a66600c646464646464a66601866e1c005200114a2264646464a666026602c004264646464646464a66602ea66602e60260182a66602e60146eb8c044c064dd5002899b8f00d375c600660326ea80145280a501533301700415333017001100214a0294052819b8848000dd69801180c1baa30023018375400a66e1cc94ccc058c048c05cdd50008a400026eb4c06cc060dd500099299980b1809180b9baa00114c0103d87a8000132330010013756603860326ea8008894ccc06c004530103d87a80001323232533301b3371e0226eb8c07000c4c060cc07cdd4000a5eb804cc014014008dd6980e001180f801180e80099198008009bab30023018375400e44a666034002298103d87a80001323232533301a300d375c60360062602e6603c6e980052f5c026600a00a0046eacc06c008c078008c0700040288c068c06c004cc88c8cc00400400c894ccc068004528099299980c19b8f375c603a00400829444cc00c00c004c074004dd6180c180c980c980c980c980c980c980c980c980a9baa300d301537540226eb8c034c054dd5001180b980c001180b00098091baa332253330123370e900218099baa001132325333017301a0021320025333014300f30153754002264646464a666036603c00426464931804801299980c1809980c9baa003132323232533301f30220021324994ccc070c05cc074dd50008991919192999811981300109924c60200062c6eb4c090004c090008c088004c078dd50008b0b181000098100011bae301e001301a37540062c2c603800260380046034002602c6ea80045858c060004c050dd50008b12999808980618091baa0011323232325333018301b002149858dd7180c800980c8011bae3017001301337540022c600860246ea800458c050004c8cc004004dd6180198089baa30093011375401a44a666026002297ae0132325333012325333013300f301437540022600c6eb8c060c054dd50008a50300c30143754601860286ea80084cc058008cc0100100044cc010010004c05c008c054004dc780291809180998098009bad30103011002375c601e002601e0046eb8c034004c8c94ccc030c03c008400458dd61806800991980080099198008009bab300e300f300f300f300f300b3754600660166ea801c894ccc03400452f5bded8c0264646464a66601c66e3d2201000021003133012337606ea4008dd3000998030030019bab300f003375c601a0046022004601e00244a666018002297ae01323332223233001001003225333012001100313233014374e660286ea4018cc050dd49bae30110013301437506eb4c0480052f5c066006006602c00460280026eb8c02c004dd598060009980180198080011807000918060008a4c26cac26644644a666014646464646464646464a666026a666026601c60286ea80304c8c94ccc060c06c0084cdc78009bae300b301737546016602e6ea805058dd7180c8009bac301830193019301930193019301930193019301537546012602a6ea80284c8c8c8c8c94ccc060cdd79807180d1baa00530153301c30153301c300e301a3754601c60346ea805d2f5c06603898103d87a80004bd7008008a5053330173375e602866036603800666036603800497ae0300d30193754600c60326ea8c034c064dd500b099b89375a600c60326ea8c018c064dd51806980c9baa01600114a06eb4c06cc070004c06c0054ccc050cdc79bae300a301637546014602c6ea8c00cc058dd51805180b1baa013488100132325333019301c0021323301b301c0023301b301c0013301b301c301d0014bd70180e0008b1bac301a001300937566006602c6ea80044c8c94ccc064c0700084c8cc06cc070008cc06cc070004cc06cc070c0740052f5c060380022c6eb0c068004c024cc020dd59801980b1baa001488100325333014300f3015375400226032602c6ea800458c94ccc05c004530103d87a8000130113301830190014bd701bac301830193019301537546012602a6ea80284004528299980919b87375a602e603000690008a99980919b8f004375c601060286ea8c004c050dd5008899b8f002375c600260286ea8c004c050dd50088a5014a04602e60300026eb8c054004c054008dd7180980099192999809180a80108008b1bac3013001300233001375660246026602660266026601e6ea8c00cc03cdd500224410022323300100100322533301300114bd6f7b630099191919299980a19b8f0070021003133018337606ea4008dd3000998030030019bab3015003375c6026004602e004602a0024646600200200444a666022002297ae01323332223233001001003225333017001100313233019374e660326ea4018cc064dd49bae30160013301937506eb4c05c0052f5c066006006603600460320026eb8c040004dd5980880099801801980a8011809800918080008a4c26cac64a66601260080022a66601860166ea8008526161533300930050011533300c300b37540042930b0b18049baa00132533300730023008375400c264646464a66601c6022004264649318030012999805980318061baa003132323232533301230150021324994ccc03cc028c040dd5000899191919299980b180c80109924c601a0062c6eb4c05c004c05c008c054004c044dd50008b0b180980098098011bae3011001300d37540062c2c601e002601e004601a00260126ea80185894ccc01cc008c020dd5000899191919299980718088010a4c2c6eb8c03c004c03c008dd7180680098049baa00116300b300837540086e1d2000370e90011ba5480015cd2ab9d5573caae7d5d02ba157441").unwrap());
+    let script_hash = compute_plutus_v2_script_hash(script.clone());
+
+    let token_a_policy = H224::from(
+        Hash::from_str("0298aa99f95e2fe0a0132a6bb794261fb7e7b0d988215da2f2de2005").unwrap(),
+    );
+    let token_a_name = AssetName::from("tokenA".to_string());
+    let token_b_policy = H224::from(
+        Hash::from_str("0298aa99f95e2fe0a0132a6bb794261fb7e7b0d988215da2f2de2005").unwrap(),
+    );
+    let token_b_name = AssetName::from("tokenB".to_string());
+    let token_b_amount: u64 = 2;
+
+    let sender_payment_hash = H224::from(
+        Hash::from_str("01e6301758a6badfab05035cffc8e3438b3aff2a4edc6544b47329c4").unwrap(),
+    );
+    let control_token_policy = script_hash;
+    let control_token_name = AssetName::from("controlToken".to_string());
+    let _mint = Some(Multiasset::from((
+        control_token_policy,
+        control_token_name.clone(),
+        -1,
+    )));
+
+    let order_datum = OrderDatum::Ok {
+        sender_payment_hash: sender_payment_hash,
+        control_token_class: AssetClass {
+            policy_id: control_token_policy,
+            asset_name: control_token_name.clone(),
+        },
+        ordered_class: AssetClass {
+            policy_id: token_b_policy,
+            asset_name: token_b_name,
+        },
+        ordered_amount: token_b_amount,
+    };
+
+    let outputs = vec![Output {
+        address: Address(hex::decode("70".to_owned() + &hex::encode(script_hash)).unwrap()),
+        value: Value::from((token_a_policy, token_a_name, 1))
+            + Value::from((control_token_policy, control_token_name.clone(), 1))
+            + Value::Coin(10),
+        datum_option: Some(Datum::from(order_datum)),
+    }];
+
+    let mut transaction = Transaction::from((Vec::new(), Vec::new()));
+    for output in outputs {
+        transaction.transaction_body.outputs.push(output.clone());
+    }
+
+    let vkeywitness = VKeyWitness {
+        vkey: hex::decode("7b155093789404780735f4501c576e9f6e2b0a486cdec70e03e1ef8b9ef99274").unwrap(),
+        signature: hex::decode("9a8b0fe674070d2813141f2cf0476be33700e36ed88b0b5d5277d620036db990eed93d815b37a68ae08c722227a4900d544d3a4b93defd3ebc5e2f0b6a078100").unwrap()
+    };
+    let sign: H224 = H224::from(
+        Hash::from_str("01e6301758a6badfab05035cffc8e3438b3aff2a4edc6544b47329c4").unwrap(),
+    );
+
+    let mint = Some(Multiasset::from((
+        control_token_policy,
+        control_token_name,
+        1,
+    )));
+    let mint_redeemer = Redeemer {
+        tag: RedeemerTag::Mint,
+        index: 0,
+        data: PlutusData::from(PallasPlutusData::Constr(Constr {
+            tag: 121,
+            any_constructor: None,
+            fields: Def([].to_vec()),
+        })),
+        ex_units: ExUnits {
+            mem: 661056,
+            steps: 159759842,
+        },
+    };
+
+    transaction.transaction_body.mint = mint;
+    transaction.transaction_body.required_signers = Some(vec![sign]);
+    transaction.transaction_witness_set.vkeywitness = Some(vec![vkeywitness]);
+    transaction.transaction_witness_set.redeemer = Some(vec![mint_redeemer]);
+    transaction.transaction_witness_set.plutus_script = Some(vec![script.clone()]);
+
+    let pallas_tx: PallasTransaction = <_>::from(transaction.clone());
+    let cbor_bytes: Vec<u8> = babbage_tx_to_cbor(&pallas_tx);
+    let mtx: ConwayMintedTx = conway_minted_tx_from_cbor(&cbor_bytes);
+
+    let redeemers = eval_phase_two(
+        &mtx,
+        &vec![],
+        None,
+        None,
+        &SlotConfig::default(),
+        false,
+        |_| (),
+    )
+    .unwrap();
+    assert_eq!(redeemers.len(), 1);
+}
+
+#[test]
+fn test_phase2_order_resolve() {
+    use core::str::FromStr;
+    use griffin_core::checks_interface::conway_minted_tx_from_cbor;
+    use griffin_core::h224::H224;
+    use griffin_core::pallas_codec::utils::MaybeIndefArray::Def;
+    use griffin_core::pallas_crypto::hash::Hash;
+    use griffin_core::pallas_primitives::conway::{
+        Constr, MintedTx as ConwayMintedTx, PlutusData as PallasPlutusData, TransactionInput,
+        TransactionOutput,
+    };
+    use griffin_core::types::{
+        compute_plutus_v2_script_hash, Address, AssetClass, AssetName, Datum, ExUnits, Input,
+        Multiasset, Output, PlutusData, PlutusScript, Redeemer, RedeemerTag, VKeyWitness, Value,
+    };
+    use griffin_core::uplc::tx::{eval_phase_two, ResolvedInput, SlotConfig};
+    use sp_core::H256;
+
+    let script = PlutusScript(hex::decode("59080c010000323232323232322253232323330063001300737540082a66600c646464646464a66601866e1c005200114a2264646464a666026602c004264646464646464a66602ea66602e60260182a66602e60146eb8c044c064dd5002899b8f00d375c600660326ea80145280a501533301700415333017001100214a0294052819b8848000dd69801180c1baa30023018375400a66e1cc94ccc058c048c05cdd50008a400026eb4c06cc060dd500099299980b1809180b9baa00114c0103d87a8000132330010013756603860326ea8008894ccc06c004530103d87a80001323232533301b3371e0226eb8c07000c4c060cc07cdd4000a5eb804cc014014008dd6980e001180f801180e80099198008009bab30023018375400e44a666034002298103d87a80001323232533301a300d375c60360062602e6603c6e980052f5c026600a00a0046eacc06c008c078008c0700040288c068c06c004cc88c8cc00400400c894ccc068004528099299980c19b8f375c603a00400829444cc00c00c004c074004dd6180c180c980c980c980c980c980c980c980c980a9baa300d301537540226eb8c034c054dd5001180b980c001180b00098091baa332253330123370e900218099baa001132325333017301a0021320025333014300f30153754002264646464a666036603c00426464931804801299980c1809980c9baa003132323232533301f30220021324994ccc070c05cc074dd50008991919192999811981300109924c60200062c6eb4c090004c090008c088004c078dd50008b0b181000098100011bae301e001301a37540062c2c603800260380046034002602c6ea80045858c060004c050dd50008b12999808980618091baa0011323232325333018301b002149858dd7180c800980c8011bae3017001301337540022c600860246ea800458c050004c8cc004004dd6180198089baa30093011375401a44a666026002297ae0132325333012325333013300f301437540022600c6eb8c060c054dd50008a50300c30143754601860286ea80084cc058008cc0100100044cc010010004c05c008c054004dc780291809180998098009bad30103011002375c601e002601e0046eb8c034004c8c94ccc030c03c008400458dd61806800991980080099198008009bab300e300f300f300f300f300b3754600660166ea801c894ccc03400452f5bded8c0264646464a66601c66e3d2201000021003133012337606ea4008dd3000998030030019bab300f003375c601a0046022004601e00244a666018002297ae01323332223233001001003225333012001100313233014374e660286ea4018cc050dd49bae30110013301437506eb4c0480052f5c066006006602c00460280026eb8c02c004dd598060009980180198080011807000918060008a4c26cac26644644a666014646464646464646464a666026a666026601c60286ea80304c8c94ccc060c06c0084cdc78009bae300b301737546016602e6ea805058dd7180c8009bac301830193019301930193019301930193019301537546012602a6ea80284c8c8c8c8c94ccc060cdd79807180d1baa00530153301c30153301c300e301a3754601c60346ea805d2f5c06603898103d87a80004bd7008008a5053330173375e602866036603800666036603800497ae0300d30193754600c60326ea8c034c064dd500b099b89375a600c60326ea8c018c064dd51806980c9baa01600114a06eb4c06cc070004c06c0054ccc050cdc79bae300a301637546014602c6ea8c00cc058dd51805180b1baa013488100132325333019301c0021323301b301c0023301b301c0013301b301c301d0014bd70180e0008b1bac301a001300937566006602c6ea80044c8c94ccc064c0700084c8cc06cc070008cc06cc070004cc06cc070c0740052f5c060380022c6eb0c068004c024cc020dd59801980b1baa001488100325333014300f3015375400226032602c6ea800458c94ccc05c004530103d87a8000130113301830190014bd701bac301830193019301537546012602a6ea80284004528299980919b87375a602e603000690008a99980919b8f004375c601060286ea8c004c050dd5008899b8f002375c600260286ea8c004c050dd50088a5014a04602e60300026eb8c054004c054008dd7180980099192999809180a80108008b1bac3013001300233001375660246026602660266026601e6ea8c00cc03cdd500224410022323300100100322533301300114bd6f7b630099191919299980a19b8f0070021003133018337606ea4008dd3000998030030019bab3015003375c6026004602e004602a0024646600200200444a666022002297ae01323332223233001001003225333017001100313233019374e660326ea4018cc064dd49bae30160013301937506eb4c05c0052f5c066006006603600460320026eb8c040004dd5980880099801801980a8011809800918080008a4c26cac64a66601260080022a66601860166ea8008526161533300930050011533300c300b37540042930b0b18049baa00132533300730023008375400c264646464a66601c6022004264649318030012999805980318061baa003132323232533301230150021324994ccc03cc028c040dd5000899191919299980b180c80109924c601a0062c6eb4c05c004c05c008c054004c044dd50008b0b180980098098011bae3011001300d37540062c2c601e002601e004601a00260126ea80185894ccc01cc008c020dd5000899191919299980718088010a4c2c6eb8c03c004c03c008dd7180680098049baa00116300b300837540086e1d2000370e90011ba5480015cd2ab9d5573caae7d5d02ba157441").unwrap());
+    let script_hash = compute_plutus_v2_script_hash(script.clone());
+
+    let token_a_policy = H224::from(
+        Hash::from_str("0298aa99f95e2fe0a0132a6bb794261fb7e7b0d988215da2f2de2005").unwrap(),
+    );
+    let token_a_name = AssetName::from("tokenA".to_string());
+    let token_b_policy = H224::from(
+        Hash::from_str("0298aa99f95e2fe0a0132a6bb794261fb7e7b0d988215da2f2de2005").unwrap(),
+    );
+    let token_b_name = AssetName::from("tokenB".to_string());
+    let token_b_amount: u64 = 2;
+
+    let sender_payment_hash = H224::from(
+        Hash::from_str("01e6301758a6badfab05035cffc8e3438b3aff2a4edc6544b47329c4").unwrap(),
+    );
+    let control_token_policy = script_hash;
+    let control_token_name = AssetName::from("controlToken".to_string());
+
+    let order_datum = OrderDatum::Ok {
+        sender_payment_hash: sender_payment_hash,
+        control_token_class: AssetClass {
+            policy_id: control_token_policy,
+            asset_name: control_token_name.clone(),
+        },
+        ordered_class: AssetClass {
+            policy_id: token_b_policy,
+            asset_name: token_b_name.clone(),
+        },
+        ordered_amount: token_b_amount,
+    };
+
+    let inputs = vec![
+        Input {
+            tx_hash: H256::from_slice(
+                hex::decode("62b64a5cca307624cd90ff002ac26c3d3faa9dec56f1f360954120ffc38306e8")
+                    .unwrap()
+                    .as_slice(),
+            ),
+            index: 0,
+        },
+        Input {
+            tx_hash: H256::from_slice(
+                hex::decode("27d67023bde9c18c75fe5939a88d3771f061d82cfa3b9bce12d1df9118b74a8d")
+                    .unwrap()
+                    .as_slice(),
+            ),
+            index: 1,
+        },
+    ];
+    let resolved_inputs = vec![
+        Output {
+            address: Address(hex::decode("70".to_owned() + &hex::encode(script_hash)).unwrap()),
+            value: Value::from((314, token_a_policy, token_a_name, 1)),
+            datum_option: Some(Datum::from(order_datum)),
+        },
+        Output {
+            address: Address(
+                hex::decode("6101e6301758a6badfab05035cffc8e3438b3aff2a4edc6544b47329c4").unwrap(),
+            ),
+            value: Value::Coin(3),
+            datum_option: None,
+        },
+    ];
+
+    let outputs = vec![Output {
+        address: Address(
+            hex::decode("6101e6301758a6badfab05035cffc8e3438b3aff2a4edc6544b47329c4").unwrap(),
+        ),
+        value: Value::from((10, token_b_policy, token_b_name, 2)),
+        datum_option: None,
+    }];
+
+    let pallas_inputs = inputs
+        .iter()
+        .map(|i| TransactionInput::from(i.clone()))
+        .collect::<Vec<_>>();
+    let pallas_resolved_inputs = resolved_inputs
+        .iter()
+        .map(|ri| TransactionOutput::from(ri.clone()))
+        .collect::<Vec<_>>();
+
+    let resolve_redeemer = Redeemer {
+        tag: RedeemerTag::Spend,
+        index: 1,
+        data: PlutusData::from(PallasPlutusData::Constr(Constr {
+            tag: 122,
+            any_constructor: None,
+            fields: Def([PallasPlutusData::Constr(Constr {
+                tag: 122,
+                any_constructor: None,
+                fields: Def([].to_vec()),
+            })]
+            .to_vec()),
+        })),
+        ex_units: ExUnits {
+            mem: 661056,
+            steps: 159759842,
+        },
+    };
+
+    let mut transaction = Transaction::from((Vec::new(), Vec::new()));
+    for input in inputs {
+        transaction.transaction_body.inputs.push(input.clone());
+    }
+    for output in outputs {
+        transaction.transaction_body.outputs.push(output.clone());
+    }
+
+    let vkeywitness = VKeyWitness {
+        vkey: hex::decode("F6E9814CE6626EB532372B1740127E153C28D643A9384F51B1B0229AEDA43717").unwrap(),
+        signature: hex::decode("b22b78b86e64a5849f51df512289c5be0c463cce6966caaecda265be3ed37b5c68c04dad7616bfa78a2e1235eb80c8b9b366c0a16a131fbcfebce8b35382b90d").unwrap()
+    };
+    let sign: H224 = H224::from(
+        Hash::from_str("01e6301758a6badfab05035cffc8e3438b3aff2a4edc6544b47329c4").unwrap(),
+    );
+
+    let mint = Some(Multiasset::from((
+        control_token_policy,
+        control_token_name,
+        -1,
+    )));
+    let burn_redeemer = Redeemer {
+        tag: RedeemerTag::Mint,
+        index: 0,
+        data: PlutusData::from(PallasPlutusData::Constr(Constr {
+            tag: 121,
+            any_constructor: None,
+            fields: Def([].to_vec()),
+        })),
+        ex_units: ExUnits {
+            mem: 661056,
+            steps: 159759842,
+        },
+    };
+
+    transaction.transaction_body.mint = mint;
+    transaction.transaction_body.required_signers = Some(vec![sign]);
+    transaction.transaction_witness_set.vkeywitness = Some(vec![vkeywitness]);
+    transaction.transaction_witness_set.redeemer = Some(vec![burn_redeemer, resolve_redeemer]);
+    transaction.transaction_witness_set.plutus_script = Some(vec![script]);
+
+    let pallas_tx: PallasTransaction = <_>::from(transaction.clone());
+    let cbor_bytes: Vec<u8> = babbage_tx_to_cbor(&pallas_tx);
+    let mtx: ConwayMintedTx = conway_minted_tx_from_cbor(&cbor_bytes);
+
+    let input_utxos: Vec<ResolvedInput> = pallas_inputs
+        .iter()
+        .zip(pallas_resolved_inputs.iter())
+        .map(|(input, output)| ResolvedInput {
+            input: input.clone(),
+            output: output.clone(),
+        })
+        .collect();
+
+    let redeemers = eval_phase_two(
+        &mtx,
+        &input_utxos,
+        None,
+        None,
+        &SlotConfig::default(),
+        false,
+        |_| (),
+    )
+    .unwrap();
+    assert_eq!(redeemers.len(), 2);
+}
+
+#[test]
+fn test_phase2_order_cancel() {
+    use core::str::FromStr;
+    use griffin_core::checks_interface::conway_minted_tx_from_cbor;
+    use griffin_core::h224::H224;
+    use griffin_core::pallas_codec::utils::MaybeIndefArray::Def;
+    use griffin_core::pallas_crypto::hash::Hash;
+    use griffin_core::pallas_primitives::conway::{
+        Constr, MintedTx as ConwayMintedTx, PlutusData as PallasPlutusData, TransactionInput,
+        TransactionOutput,
+    };
+    use griffin_core::types::{
+        compute_plutus_v2_script_hash, Address, AssetClass, AssetName, Datum, ExUnits, Input,
+        Multiasset, Output, PlutusData, PlutusScript, Redeemer, RedeemerTag, VKeyWitness, Value,
+    };
+    use griffin_core::uplc::tx::{eval_phase_two, ResolvedInput, SlotConfig};
+    use sp_core::H256;
+
+    let script = PlutusScript(hex::decode("59080c010000323232323232322253232323330063001300737540082a66600c646464646464a66601866e1c005200114a2264646464a666026602c004264646464646464a66602ea66602e60260182a66602e60146eb8c044c064dd5002899b8f00d375c600660326ea80145280a501533301700415333017001100214a0294052819b8848000dd69801180c1baa30023018375400a66e1cc94ccc058c048c05cdd50008a400026eb4c06cc060dd500099299980b1809180b9baa00114c0103d87a8000132330010013756603860326ea8008894ccc06c004530103d87a80001323232533301b3371e0226eb8c07000c4c060cc07cdd4000a5eb804cc014014008dd6980e001180f801180e80099198008009bab30023018375400e44a666034002298103d87a80001323232533301a300d375c60360062602e6603c6e980052f5c026600a00a0046eacc06c008c078008c0700040288c068c06c004cc88c8cc00400400c894ccc068004528099299980c19b8f375c603a00400829444cc00c00c004c074004dd6180c180c980c980c980c980c980c980c980c980a9baa300d301537540226eb8c034c054dd5001180b980c001180b00098091baa332253330123370e900218099baa001132325333017301a0021320025333014300f30153754002264646464a666036603c00426464931804801299980c1809980c9baa003132323232533301f30220021324994ccc070c05cc074dd50008991919192999811981300109924c60200062c6eb4c090004c090008c088004c078dd50008b0b181000098100011bae301e001301a37540062c2c603800260380046034002602c6ea80045858c060004c050dd50008b12999808980618091baa0011323232325333018301b002149858dd7180c800980c8011bae3017001301337540022c600860246ea800458c050004c8cc004004dd6180198089baa30093011375401a44a666026002297ae0132325333012325333013300f301437540022600c6eb8c060c054dd50008a50300c30143754601860286ea80084cc058008cc0100100044cc010010004c05c008c054004dc780291809180998098009bad30103011002375c601e002601e0046eb8c034004c8c94ccc030c03c008400458dd61806800991980080099198008009bab300e300f300f300f300f300b3754600660166ea801c894ccc03400452f5bded8c0264646464a66601c66e3d2201000021003133012337606ea4008dd3000998030030019bab300f003375c601a0046022004601e00244a666018002297ae01323332223233001001003225333012001100313233014374e660286ea4018cc050dd49bae30110013301437506eb4c0480052f5c066006006602c00460280026eb8c02c004dd598060009980180198080011807000918060008a4c26cac26644644a666014646464646464646464a666026a666026601c60286ea80304c8c94ccc060c06c0084cdc78009bae300b301737546016602e6ea805058dd7180c8009bac301830193019301930193019301930193019301537546012602a6ea80284c8c8c8c8c94ccc060cdd79807180d1baa00530153301c30153301c300e301a3754601c60346ea805d2f5c06603898103d87a80004bd7008008a5053330173375e602866036603800666036603800497ae0300d30193754600c60326ea8c034c064dd500b099b89375a600c60326ea8c018c064dd51806980c9baa01600114a06eb4c06cc070004c06c0054ccc050cdc79bae300a301637546014602c6ea8c00cc058dd51805180b1baa013488100132325333019301c0021323301b301c0023301b301c0013301b301c301d0014bd70180e0008b1bac301a001300937566006602c6ea80044c8c94ccc064c0700084c8cc06cc070008cc06cc070004cc06cc070c0740052f5c060380022c6eb0c068004c024cc020dd59801980b1baa001488100325333014300f3015375400226032602c6ea800458c94ccc05c004530103d87a8000130113301830190014bd701bac301830193019301537546012602a6ea80284004528299980919b87375a602e603000690008a99980919b8f004375c601060286ea8c004c050dd5008899b8f002375c600260286ea8c004c050dd50088a5014a04602e60300026eb8c054004c054008dd7180980099192999809180a80108008b1bac3013001300233001375660246026602660266026601e6ea8c00cc03cdd500224410022323300100100322533301300114bd6f7b630099191919299980a19b8f0070021003133018337606ea4008dd3000998030030019bab3015003375c6026004602e004602a0024646600200200444a666022002297ae01323332223233001001003225333017001100313233019374e660326ea4018cc064dd49bae30160013301937506eb4c05c0052f5c066006006603600460320026eb8c040004dd5980880099801801980a8011809800918080008a4c26cac64a66601260080022a66601860166ea8008526161533300930050011533300c300b37540042930b0b18049baa00132533300730023008375400c264646464a66601c6022004264649318030012999805980318061baa003132323232533301230150021324994ccc03cc028c040dd5000899191919299980b180c80109924c601a0062c6eb4c05c004c05c008c054004c044dd50008b0b180980098098011bae3011001300d37540062c2c601e002601e004601a00260126ea80185894ccc01cc008c020dd5000899191919299980718088010a4c2c6eb8c03c004c03c008dd7180680098049baa00116300b300837540086e1d2000370e90011ba5480015cd2ab9d5573caae7d5d02ba157441").unwrap());
+    let script_hash = compute_plutus_v2_script_hash(script.clone());
+
+    let token_b_policy = H224::from(
+        Hash::from_str("0298aa99f95e2fe0a0132a6bb794261fb7e7b0d988215da2f2de2005").unwrap(),
+    );
+    let token_b_name = AssetName::from("tokenB".to_string());
+    let token_a_policy = token_b_policy;
+    let token_a_name = AssetName::from("tokenA".to_string());
+    let token_b_amount: u64 = 2;
+
+    let sender_payment_hash = H224::from(
+        Hash::from_str("5b6de1be218ebb35fc08b2983e3a1d72aec969c8d2a6301212e2ea9a").unwrap(),
+    );
+    let control_token_policy = script_hash;
+    let control_token_name = AssetName::from("controlToken".to_string());
+
+    let order_datum = OrderDatum::Ok {
+        sender_payment_hash: sender_payment_hash,
+        control_token_class: AssetClass {
+            policy_id: control_token_policy,
+            asset_name: control_token_name.clone(),
+        },
+        ordered_class: AssetClass {
+            policy_id: token_b_policy,
+            asset_name: token_b_name,
+        },
+        ordered_amount: token_b_amount,
+    };
+
+    let inputs = vec![Input {
+        tx_hash: H256::from_slice(
+            hex::decode("88832d2909740fdedac6b39348303b62a2d3d7f6d25a79c349768fe113dab451")
+                .unwrap()
+                .as_slice(),
+        ),
+        index: 0,
+    }];
+    let resolved_inputs = vec![Output {
+        address: Address(hex::decode("70".to_owned() + &hex::encode(script_hash)).unwrap()),
+        value: Value::from((314, token_a_policy, token_a_name, 1))
+            + Value::from((control_token_policy, control_token_name.clone(), 1)),
+        datum_option: Some(Datum::from(order_datum)),
+    }];
+
+    let pallas_inputs = inputs
+        .iter()
+        .map(|i| TransactionInput::from(i.clone()))
+        .collect::<Vec<_>>();
+    let pallas_resolved_inputs = resolved_inputs
+        .iter()
+        .map(|ri| TransactionOutput::from(ri.clone()))
+        .collect::<Vec<_>>();
+
+    let cancel_redeemer = Redeemer {
+        tag: RedeemerTag::Spend,
+        index: 0,
+        data: PlutusData::from(PallasPlutusData::Constr(Constr {
+            tag: 122,
+            any_constructor: None,
+            fields: Def([PallasPlutusData::Constr(Constr {
+                tag: 121,
+                any_constructor: None,
+                fields: Def([].to_vec()),
+            })]
+            .to_vec()),
+        })),
+        ex_units: ExUnits {
+            mem: 661056,
+            steps: 159759842,
+        },
+    };
+
+    let mut transaction = Transaction::from((Vec::new(), Vec::new()));
+    for input in inputs {
+        transaction.transaction_body.inputs.push(input.clone());
+    }
+
+    let burn_redeemer = Redeemer {
+        tag: RedeemerTag::Mint,
+        index: 0,
+        data: PlutusData::from(PallasPlutusData::Constr(Constr {
+            tag: 121,
+            any_constructor: None,
+            fields: Def([].to_vec()),
+        })),
+        ex_units: ExUnits {
+            mem: 661056,
+            steps: 159759842,
+        },
+    };
+    let mint = Some(Multiasset::from((
+        control_token_policy,
+        control_token_name,
+        -1,
+    )));
+
+    transaction.transaction_body.mint = mint;
+    transaction.transaction_body.required_signers = Some(vec![sender_payment_hash]);
+    let vkeywitness = VKeyWitness {
+        vkey: hex::decode("F6E9814CE6626EB532372B1740127E153C28D643A9384F51B1B0229AEDA43717").unwrap(),
+        signature: hex::decode("A4ACDA77397F7A80B21FA17AE95FCC99C255069B8135897BA8A7A5EC0E829DBA91171FBF794C1A5E6249263B04075C659BDEBA1B1E10E38F734539626BFF6905").unwrap()
+    };
+    transaction.transaction_witness_set.vkeywitness = Some(vec![vkeywitness]);
+    transaction.transaction_witness_set.redeemer = Some(vec![cancel_redeemer, burn_redeemer]);
+    transaction.transaction_witness_set.plutus_script = Some(vec![script]);
+
+    let pallas_tx: PallasTransaction = <_>::from(transaction.clone());
+    let cbor_bytes: Vec<u8> = babbage_tx_to_cbor(&pallas_tx);
+    let mtx: ConwayMintedTx = conway_minted_tx_from_cbor(&cbor_bytes);
+
+    let input_utxos: Vec<ResolvedInput> = pallas_inputs
+        .iter()
+        .zip(pallas_resolved_inputs.iter())
+        .map(|(input, output)| ResolvedInput {
+            input: input.clone(),
+            output: output.clone(),
+        })
+        .collect();
+
+    let redeemers = eval_phase_two(
+        &mtx,
+        &input_utxos,
+        None,
+        None,
+        &SlotConfig::default(),
+        false,
+        |_| (),
+    )
+    .unwrap();
+    assert_eq!(redeemers.len(), 2);
+}
